[refactor] 게시글 상세 페이지 클라이언트 분리
