+++ src/app/(main)/community/[id]/page.tsx
@@ -1,439 +1,55 @@
-'use client';
-
-import { useQueryClient } from '@tanstack/react-query';
-import { use, useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
+// 서버 컴포넌트
+import PostDetailClient from '@/components/community/PostDetailClient';
+import { createServerClient } from '@supabase/ssr';
+import { cookies } from 'next/headers';
 import {
   getPost,
   getComments,
-  createComment,
-  deletePost,
-  deleteComment,
   incrementViewCount,
 } from '@/services/communityService';
-import { reportPost, ReportReason } from '@/services/reportService';
-import { supabase } from '@/lib/supabase';
-import type { Post, Comment } from '@/types/community';
-import Image from 'next/image';
-import LoadingSpinner from '@/components/common/LoadingSpinner';
-import { useAuth } from '@/hooks/useAuth';
-import { useLike } from '@/hooks/useLike';
-import { showErrorToast, showSuccessToast } from '@/lib/toast';
-import ReportModal from '@/components/community/ReportModal';
-import PostDetailCard, {
-  PostDetailCardData,
-} from '@/components/community/PostDetailCard';
-import { timeAgo } from '@/utils/timeAgo';
+import { notFound } from 'next/navigation';
 
-export default function PostDetailPage({
+async function getInitialData(id: string) {
+  const cookieStore = await cookies();
+  const supabase = createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    { cookies: { getAll: () => cookieStore.getAll() } },
+  );
+
+  const [
+    post,
+    comments,
+    {
+      data: { user },
+    },
+  ] = await Promise.all([
+    getPost(id),
+    getComments(id),
+    supabase.auth.getUser(),
+  ]);
+
+  if (!post) notFound();
+
+  await incrementViewCount(id);
+
+  return { post, comments, userId: user?.id ?? null };
+}
+
+export default async function PostDetailPage({
   params,
 }: {
   params: Promise<{ id: string }>;
 }) {
-  const { id } = use(params);
-  const router = useRouter();
-  const [post, setPost] = useState<Post | null>(null);
-  const [comments, setComments] = useState<Comment[]>([]);
-  const [comment, setComment] = useState('');
-  const [userId, setUserId] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [editingCommentId, setEditingCommentId] = useState<string | null>(null);
-  const [editingContent, setEditingContent] = useState('');
-  const { user } = useAuth();
-  const { likeCount, isLiked, toggle } = useLike(id, user?.id ?? '');
-  const queryClient = useQueryClient();
-
-  const [reportModalOpen, setReportModalOpen] = useState(false);
-
-  useEffect(() => {
-    const load = async () => {
-      try {
-        const [
-          postData,
-          commentsData,
-          {
-            data: { user },
-          },
-        ] = await Promise.all([
-          getPost(id),
-          getComments(id),
-          supabase.auth.getUser(),
-        ]);
-        setPost(postData);
-        setComments(commentsData);
-        setUserId(user?.id ?? null);
-        await incrementViewCount(id);
-      } catch (e) {
-        console.error(e);
-      } finally {
-        setLoading(false);
-      }
-    };
-    load();
-  }, [id]);
-
-  const handleCommentSubmit = async () => {
-    if (!userId) {
-      showErrorToast('로그인이 필요합니다.');
-      return;
-    }
-    if (!comment.trim()) {
-      showErrorToast('댓글을 입력해주세요.');
-      return;
-    }
-    try {
-      const newComment = await createComment({
-        post_id: id,
-        user_id: userId,
-        content: comment,
-      });
-      setComments((prev) => [...prev, newComment]);
-      setComment('');
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  const handleDeletePost = async () => {
-    if (!confirm('게시글을 삭제하시겠습니까?')) return;
-    try {
-      await deletePost(id);
-      queryClient.invalidateQueries({ queryKey: ['posts'] });
-      showSuccessToast('게시글이 삭제되었습니다.', '🗑️');
-      router.push('/community');
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  const handleEditComment = async (commentId: string) => {
-    if (!editingContent.trim()) return;
-    try {
-      await supabase
-        .from('comments')
-        .update({ content: editingContent })
-        .eq('id', commentId);
-      setComments((prev) =>
-        prev.map((c) =>
-          c.id === commentId ? { ...c, content: editingContent } : c,
-        ),
-      );
-      setEditingCommentId(null);
-      setEditingContent('');
-      showSuccessToast('댓글이 수정되었습니다.', '✅');
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  const handleDeleteComment = async (commentId: string) => {
-    if (!confirm('댓글을 삭제하시겠습니까?')) return;
-    try {
-      await deleteComment(commentId);
-      setComments((prev) => prev.filter((c) => c.id !== commentId));
-      showSuccessToast('댓글이 삭제되었습니다.', '🗑️');
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  // ── 신고 제출 핸들러 ──
-  const handleReportSubmit = async (reason: ReportReason) => {
-    if (!userId) {
-      showErrorToast('로그인이 필요합니다.');
-      return;
-    }
-    try {
-      await reportPost(userId, id, reason);
-      showSuccessToast('신고가 접수되었습니다.', '🚨');
-      setReportModalOpen(false);
-    } catch (e: unknown) {
-      if (e instanceof Error && e.message === 'ALREADY_REPORTED') {
-        showErrorToast('이미 신고한 게시물입니다.');
-        setReportModalOpen(false);
-      } else {
-        showErrorToast('신고 처리 중 오류가 발생했습니다.');
-      }
-    }
-  };
-
-  if (loading)
-    return (
-      <div className="flex justify-center p-20">
-        <LoadingSpinner />
-      </div>
-    );
-  if (!post)
-    return (
-      <div className="flex justify-center p-20 text-gray-400">
-        게시글을 찾을 수 없습니다.
-      </div>
-    );
-
-  const isOwner = userId === post.user_id;
-
-  const handleShare = async () => {
-    const url = window.location.href;
-    const title = post?.title ?? '';
-
-    if (navigator.share && /Mobi|Android/i.test(navigator.userAgent)) {
-      try {
-        await navigator.share({ title, url });
-      } catch {
-        // 취소 무시
-      }
-    } else {
-      try {
-        await navigator.clipboard.writeText(url);
-        showSuccessToast('링크가 복사되었습니다!', '🔗');
-      } catch {
-        showErrorToast('복사에 실패했습니다.');
-      }
-    }
-  };
-
-  const postDetailCardData: PostDetailCardData = {
-    nickname: post.nickname ?? '알 수 없음',
-    avatar_url: post.avatar_url,
-    role: post.role,
-    created_at: post.created_at,
-    title: post.title,
-    image_url: post.image_url,
-    content: post.content,
-    likeCount: likeCount ?? 0,
-    commentCount: comments.length,
-    view_count: post.view_count,
-  };
+  const { id } = await params;
+  const { post, comments, userId } = await getInitialData(id);
 
   return (
-    <div className="max-w-2xl mx-auto p-4 space-y-4">
-      {/* 뒤로가기 */}
-      <button
-        onClick={() => router.push(`/community`)}
-        className="flex items-center gap-2 px-2.5 py-2 border-2 border-white bg-white text-black text-sm font-medium rounded-xl hover:bg-(--color-btn-focus) hover:text-white transition-colors duration-200 cursor-pointer"
-      >
-        <svg
-          width="16"
-          height="16"
-          viewBox="0 0 16 16"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M10 12L6 8L10 4"
-            stroke="currentColor"
-            strokeWidth="1.5"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </svg>
-        목록으로
-      </button>
-
-      {/* ── 게시글 카드 ── */}
-      <PostDetailCard
-        post={postDetailCardData}
-        isLiked={isLiked}
-        onLike={() => {
-          if (!userId) {
-            showErrorToast('로그인이 필요합니다.');
-            return;
-          }
-          toggle();
-        }}
-        headerActions={
-          <>
-            {isOwner ? (
-              <>
-                <button
-                  title="수정하기"
-                  onClick={() => router.push(`/community/${id}/edit`)}
-                >
-                  <Image src="/postEdit.svg" alt="" width={30} height={30} />
-                </button>
-                <button title="삭제하기" onClick={handleDeletePost}>
-                  <Image src="/postDelete.svg" alt="" width={32} height={32} />
-                </button>
-              </>
-            ) : (
-              <button
-                title="신고하기"
-                onClick={() => setReportModalOpen(true)}
-                className="w-8 h-8 flex items-center justify-center"
-              >
-                <Image src="/postReport.svg" alt="" width={27} height={27} />
-              </button>
-            )}
-            <button
-              title="공유하기"
-              onClick={handleShare}
-              className="w-8 h-8 flex items-center justify-center"
-            >
-              <Image src="/postShare.svg" alt="" width={18} height={18} />
-            </button>
-          </>
-        }
-      />
-
-      {/* ── 댓글 카드 ── */}
-      <div className="bg-white rounded-2xl shadow-sm border border-gray-100 overflow-hidden">
-        <div className="px-5 pt-5 pb-3 flex items-center gap-2 border-b border-gray-100">
-          <Image
-            src="/postComment.svg"
-            alt="댓글"
-            width={16}
-            height={16}
-            className="opacity-40"
-          />
-          <h2 className="text-sm font-semibold text-gray-800">댓글</h2>
-        </div>
-
-        {/* 댓글 입력 */}
-        <div className="px-5 py-4">
-          <div className="flex items-center gap-2">
-            <label htmlFor="comment-input" className="sr-only">
-              댓글 입력
-            </label>
-            <input
-              type="text"
-              id="comment-input"
-              placeholder="댓글을 입력하세요..."
-              value={comment}
-              onChange={(e) => setComment(e.target.value)}
-              onKeyDown={(e) => e.key === 'Enter' && handleCommentSubmit()}
-              className="flex-1 bg-gray-50 border border-gray-200 rounded-xl px-4 py-2.5 text-sm outline-none focus:border-gray-400 transition-colors placeholder:text-gray-400"
-            />
-            <button
-              onClick={handleCommentSubmit}
-              className="w-10 h-10 flex items-center justify-center shrink-0 transition-colors cursor-pointer"
-            >
-              <Image
-                src="/postCommentSubmit.svg"
-                alt="전송"
-                width={30}
-                height={30}
-              />
-            </button>
-          </div>
-        </div>
-
-        {/* 댓글 목록 */}
-        <div className="px-5 pb-5 space-y-4">
-          {comments.map((c, index) => (
-            <div key={c.id}>
-              {index > 0 && <div className="border-t border-gray-50 mb-4" />}
-              <div className="flex gap-3">
-                <div className="w-8 h-8 rounded-full bg-gray-200 shrink-0 overflow-hidden">
-                  {c.avatar_url && (
-                    <Image
-                      src={c.avatar_url}
-                      alt={`${c.nickname} 프로필 이미지`}
-                      width={32}
-                      height={32}
-                      className="w-full h-full object-cover"
-                    />
-                  )}
-                </div>
-                <div className="flex-1 min-w-0">
-                  <div className="flex items-center gap-1.5 mb-1">
-                    <span className="text-sm font-semibold text-gray-900">
-                      {c.nickname}
-                    </span>
-                    {c.belt_level && (
-                      <span className="text-xs text-gray-400">
-                        {c.belt_level}
-                      </span>
-                    )}
-                  </div>
-                  {editingCommentId === c.id ? (
-                    <div className="flex items-center gap-2 mt-1">
-                      <label
-                        htmlFor={`edit-comment-${c.id}`}
-                        className="sr-only"
-                      >
-                        댓글 수정 입력
-                      </label>
-                      <input
-                        id={`edit-comment-${c.id}`}
-                        type="text"
-                        value={editingContent}
-                        onChange={(e) => setEditingContent(e.target.value)}
-                        onKeyDown={(e) =>
-                          e.key === 'Enter' && handleEditComment(c.id)
-                        }
-                        className="flex-1 bg-gray-50 border border-gray-200 rounded-xl px-3 py-1.5 text-sm outline-none focus:border-gray-400"
-                        autoFocus
-                      />
-                      <button
-                        onClick={() => handleEditComment(c.id)}
-                        className="text-xs text-blue-500 font-medium"
-                      >
-                        저장
-                      </button>
-                      <button
-                        onClick={() => setEditingCommentId(null)}
-                        className="text-xs text-gray-400"
-                      >
-                        취소
-                      </button>
-                    </div>
-                  ) : (
-                    <p className="text-sm text-gray-700 leading-relaxed">
-                      {c.content}
-                    </p>
-                  )}
-                  <div className="flex items-center gap-3 mt-1.5">
-                    <span className="flex items-center gap-1 text-xs text-gray-400">
-                      <Image
-                        src="/postTime.svg"
-                        alt=""
-                        aria-hidden="true"
-                        width={11}
-                        height={11}
-                        className="opacity-50"
-                      />
-                      {timeAgo(c.created_at)}
-                    </span>
-                    {userId === c.user_id && (
-                      <>
-                        <button
-                          onClick={() => {
-                            setEditingCommentId(c.id);
-                            setEditingContent(c.content);
-                          }}
-                          aria-label={`${c.nickname}의 댓글 수정`}
-                          className="text-xs text-gray-400 hover:text-gray-600 transition-colors"
-                        >
-                          수정
-                        </button>
-                        <button
-                          onClick={() => handleDeleteComment(c.id)}
-                          aria-label={`${c.nickname}의 댓글 삭제`}
-                          className="flex items-center gap-1 text-xs text-red-400 hover:text-red-600 transition-colors"
-                        >
-                          삭제
-                        </button>
-                      </>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </div>
-          ))}
-          {comments.length === 0 && (
-            <p className="text-center text-sm text-gray-400 py-4">
-              첫 번째 댓글을 남겨보세요!
-            </p>
-          )}
-        </div>
-      </div>
-
-      {/* ── 신고 모달 ── */}
-      <ReportModal
-        isOpen={reportModalOpen}
-        onClose={() => setReportModalOpen(false)}
-        onSubmit={handleReportSubmit}
-      />
-    </div>
+    <PostDetailClient
+      id={id}
+      initialPost={post}
+      initialComments={comments}
+      userId={userId}
+    />
   );
 }
